♻️ refactor(manifolds): make the Lorentzian precondition a type, not a runtime scan
`causal_character`, `proper_time` and `proper_distance` dispatched on
`AbstractChart` and then scanned `metric.signature`, so
`causal_character(cart3d, a, b)` *resolved* and only then refused. They
advertised a generality they do not have -- of the seven metrics in the library,
exactly one is Lorentzian.

Three changes, none to the public namespace.

**The precondition is now a type.** `AbstractLorentzianMetricField` joins
`AbstractDiagonalMetricField` as a structural marker on the metric hierarchy --
one is about the *shape* of the matrix, the other about its *signature*, and
`MinkowskiMetric` is both. The three verbs dispatch on it, so a Riemannian metric
has no method rather than a method that accepts and rejects. `_require_lorentzian`
is deleted.

This is also what makes the set extensible. A curved spacetime metric
(Schwarzschild, FLRW) inherits the marker and acquires all three verbs; before,
each verb would have had to learn about it.

**The implementation moved to `_src/minkowski/causality.py`**, matching the house
style already used by `scale_factors` and `metric_matrix`: a verb is declared
once in `coordinaxs.api`, and each manifold registers its own dispatch in its own
package.

**`interval` stays in `_src/manifolds/`** because it genuinely is generic -- the
signed quadratic form is defined for every metric, equals `separation**2` when
Riemannian, and is what the other three read. `angle_between` stays too: since
#687 it handles spacelike pairs under an indefinite metric, so it works on
Euclidean, spherical *and* Minkowski. The narrow set is exactly these three.

The public spelling is unchanged -- `cxm.causal_character(chart, a, b)` still
works, via a chart-level overload that resolves the metric and redispatches,
mirroring how `norm` and `separation` layer chart-level over metric-level.

A fallback overload on `AbstractMetricField` turns plum's `NotFoundLookupError`
-- accurate, but a resolution dump with the call echoed back -- into a sentence
naming the requirement and pointing at `interval` or `separation`.
